Use Chessie for error propagation.
Chessie is basically an F# implementation of Haskell's Either monad
(railway-oriented programming, see
 http://fsharpforfunandprofit.com/posts/recipe-part2/).

I'm mainly adding it in here to clean up the mess of result/error
types in the conversion stage.  It might be removed again later if
we find a better way of doing this.

Chessie is public domain: see
https://github.com/fsprojects/Chessie/blob/master/LICENSE.txt