Massive shakeup of variables.
We now check the variable environment on ViewDefs to see if there
are any duplicates both inside the def, and also with regards to
the global variables.

This includes changes that make duplicate variables in ViewDefs
impossible, which has caused the ticketed lock example to have to
change to follow suit.

We still don't yet use the results from this to check and infer
types, but this is coming up next.