Use partial models for creating expressions.
This will make it much easier to slot in the variable lookup
tomorrow.

I've also added a ‘this takes any expression in’ active pattern,
but this isn't yet used.  This will be key to making sure equalities
can take either form of variable in.

Tests are currently being held up with duct tape.  Will fix later.