Make the global adder sound.
Sound for Z3, at least -- HSF will likely need work.

We now translate the empty view into a func-view called 'emp', so
that we can give emp parameters.  This is a bit of a hack, but
for now it works (modulo the parser needing to be told not to
allow emp!).