Rename and reorganise model types.
The new names and types are more consistent: we always talk about a View
of Funcs, and make a distinction between a Term and an Axiom.  The type
of an entry in D is always ViewDef, ie a map from DView to BoolExpr.

This has also required severe cleanup and surgery in the pretty printer.