Make flatten collapse views into funcs.
This is a movement of the logic previously in Horn, and means that the
end result of the model translation stage is a model whose viewdefs are
on DFuncs and whose terms have GView weakest preconditions and VFunc
goals.

This rather drastic change seems to work, for now at least.