Add Peterson's algorithm proof.
Finally, after 10,000 years, it looks like we can prove this.  Sadly, I
had to write this out by hand, because trying to ask HSF to synthesise
me a proof not only kept failing due to me not getting the right views,
but also made HSF insanely slow and prone to
out-of-memory-errors.  (Needs more optimisation?)

This balloons the regression test time :(