Slightly change and move reader/writer example.
This still had the ticketed lock in it, which was ok in Starling
before I fixed #16, but now fails because of missing views.
I've chopped out the ticketed lock code.

This also currently fails, so I've moved it to WIP so it doesn't
get tangled in the regression test results.

Hope nobody minds!