Implement conservative id elimination.
Any two nodes with the same view and only id edges travelling between
them are unified with the edges deleted.

This is conservative in that it only runs through the nodes once, and
doesn't update its nodes-to-visit set to account for prior
unifications.  To what extent either is needed is not known yet.

This manages to get Peterson's algorithm from 416 proof terms to
364, a saving of one-eighth.

This is entirely a name-based optimisation: if Id is somehow defined in
the semantics to be anything other than true, it will be unsound.  If
the id transition is not named Id, it will be incomplete.

I haven't yet treble-checked this for soundness, and it needs some unit
tests writing, but it passes the regression tests.