Convert Graph and optimisation to double-adjlists.
Mike pointed out that, if we want to know the outs and ins of a
node, the easiest thing to do is store _two_ adjacency lists and
keep them in sync.  So, that's what we're now doing.

There is a lot of cruft in here that needs to be cleaned up.
There also seems to be a heavy performance regression in
this code.