Implement multi-command runs in the frontend.
So, one can now write

{|A()|}
tb = ta;
tc = tb;
{|B()|}

and the result will be that Starling will consider that entire run
as one single command.

Skip is no longer a special case in the AST (though it _is_ in parsing):

{|A()|}
;
{|B()|}

is parsed internally as the empty command list.

This has also pulled in a lot of frontend refactoring, mostly to do with
internal representations of AST nodes and errors.

Of this, there is one bit (merging atomics and locals) that I'm
considering reverting.  This is because the parse

{|A()|}
<sa = tb>; <ta = sa>;
{|B()|}

incorrectly suggests that these are two separate atomic actions, and

{|A()|}
<sa = tb; ta = sa>;
{|B()|}

is difficult to implement as it stands.  This merge was a bit of
over-eagerness on my part!

I'm not sure what people think, so I'll leave this for now.