Add Matt P's local view remover.
This is currently just limited to cases with one in-node and one
out-node.

To enable, pass -Ox.  I'll be prefixing all experimental optimisations
with x in future.

I might add some syntax to views to allow the user to specify whether
they care if the view is elided in the proof.  If this is made explicit,
this optimisation can be enabled by default.