Fix and generalise ITE optimisation.
This was being applied over-liberally, as there were no checks to
make sure the views were correct.

At the same time, it wasn't being applied generally enough, as it
wasn't dealing with the possibility of the other side of the ITE
being emp.

Examples for this change include Fail/ticketLockFlippedLoop for
the soundness issue, and Pass/spinLock for the completeness issue.