Shuffle the new optimisation xforms a bit.
I added a few new transformation types while this was being made,
so I've ported it to use those.

More controversially perhaps, I've changed this to a seq expression.
Not sure if I'm being overly wary of mutability, but this seemed a
bit cleaner anyway?

I'd have runTransforms take a seq anyway, but it uses foldFastTerm,
which is based on sequences.