Move unknown-view desugaring to modeller phase.
This allows {|?|} to be expanded to {|n(locals)?|} _before_ we
expand 'search n', which means the latter will include the former!

The upshot of this is that we can now write elided-views proofs.
I've added an example of using this for Peterson, as well as a Z3
proof unmangled by hand from HSF's output.

Phew.  I think this is it for this branch.