Refactor reification to not use powerbag
Altered to use ordered views in multiple places to account for the
order of matching.

Reify now uses the definitions to build the patterns, rather than
filtering the powerbag of the view.

Correct behaviour for matching with a definitions that has a definition
that does is not symettric about repeated views in the constraint.
We need to add a test for this case.