Finish initial implementation of symbolic variables.
Variables in (most) expressions can now be specified as a
symbolic expression binding zero or more additional expressions.
Starling makes sure that no such symbolic variables hit backends.

This is not yet hooked up to the syntax, is missing unit tests,
and needs a LOT of refactoring.

This monster took three days to finish.