Update AST with positional information.
Now Expression and Command have both been wrapped around a Node<t> type.
This happens via the `nodify` combinator in Parser which has signature
Parser<t, u> -> Parser(Node<t>, u> and wraps up the original in
positional information.

n.b. This changes the structure of the AST quite a lot and so the tests
are broken for this commit and will not type check.