Include type information in Results
Updated Command representation to make Results be a list of Param's,
this aligns the Command representation with the semantic representation
as well as allowing better matching of results to semantics based on
type.

Mixed in with this commit are some fixes to the coreSemantics and
Modeller that were causing some issues when running ./regress.sh which
should be fixed now with this change.