Fix intermediates in frame
The frame used to only generate Before/After pairs which fails when the
order of framing is swapped since now the frame has to worry about a
level of intermediates. The frame operates as before, looking for
unmatched After variables and then applying a Before but this time it
looks for intermediates on the variable too and tries to apply the final
intermediate.