Add view preprocessing step.
This is intended to 'pre-expand' possible equalities between
iterated funcs so that the pattern matcher doesn't need to.

For example, the iterated view

(G1 -> A(a)[i]) * (G2 -> A(b)[j])

will now hopefully preprocess to

(G1 -> A(a)[i])
* (G2 -> A(b)[j])
* (G1 && G2 && a == b -> A(a)[i+j])

Of course, it hasn't been tested yet, but at least the non-iterated
case goes through regress.