Add syntax for array accesses.
This has required a lot of invasive surgery throughout Starling,
as previously everything expected a lvalue to be a string.  It can
now be an array!

The array stuff still isn't hooked up to modelling yet, so any use
of arrays will fail miserably.

This also fixes some AST weirdness for expressions.