Change symbolic syntax.
Instead of the previous syntax

```
%{ #1 := #2.lock }(x, y)
```

We now have the rather more readable

```
%{ [|x|] := [|y|].lock }
```

This has required a fair amount of surgery to Starling, as the
symbolic representation has changed from sentence/arguments to a
list of inline strings and expressions.

All of the currently passing GRASShopper examples have been
converted over.