Unify local and atomic primitive command sets.
This rather large refactor means that the only difference between assignments
etc. in <| atomic brackets |> and outside them is scoping: atomic bracket
commands can use shared variables.

This also, in principle, means that CAS and ++/-- are available for use
outside of atomic brackets.  This has not been rigorously tested yet.