(Hopefully) fix iterator nontermination
As the postcondition is always a known constant iterator, we can
just consider it as a finite number of iterator-1 views, and, by
subtracting each individually, we can replace the postcondition
remainder view with an added check to the guard of the
postcondition failure view.

We can also eliminate iterated views from a large amount of the
tool frontend, simplifying things a lot.

Thanks to Matt Parkinson for the suggested fix!