Add nondeterministic choice and Go-style if
Instead of `if (foo) { bar } else { baz }`, one may now write
`if foo { bar } else { baz }`, as in the Go language.

One may also now write `if * { bar } [ else { baz } ]` to simulate a
nondeterministic choice.  Without an `else`, this models a piece of code
that may or may not run; with an `else`, it models a choice between two
alternatives.